fix(demo): fix Long wrapper identity comparison in demo service
Fix subtle Java Long comparison bug.

The demo portfolio service was using == for Long wrapper
comparison instead of .equals(), causing false negatives
for values outside the Long cache range (-128 to 127).

Replaces reference equality with value equality check.

+++ computation/src/main/java/com/tickonomics/computation/demo/VirtualPortfolio.java
@@ -62,7 +62,7 @@ public class VirtualPortfolio {
 
   public VirtualPortfolioTrade closePosition(long positionId, double exitPrice) {
     VirtualPortfolioPosition position = positionRepository.findOpenPositions().stream()
-        .filter(p -> p.id() == positionId)
+        .filter(p -> p.id().longValue() == positionId)
         .findFirst()
         .orElseThrow(() -> new IllegalArgumentException("No open position with id: " + positionId));
